Move `noris.types` and `noris.queries` to `utils` folder and update related imports.
